// web/src/gen/model/lookupContentItemsParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type LookupContentItemsParams = {
  /**
   * @nullable
   */
  contentTypeId?: string | null;
  /**
   * @nullable
   */
  contentTypeKey?: string | null;
  /**
   * @nullable
   */
  keyword?: string | null;
  limit?: number;
};